+++ ReleaseNotes.md
@@ -19,6 +19,7 @@
 
 * Upgraded Serenity BDD to latest 3.6.20
 * Added first version of FAQ.md
+* TGR-867: ExternalUrl server uses source URL as default health check URL if not set
 
 -------
 * Serenity BDD 3.3.10